Clean some leftover tails from the #1997.
- move port related funcs and constants to its own file. register.go
  sounded wrong for that
- update the SKS code and tests to deal with the change
- also move a new function there that will be used for probing quite
  often
